Move Enums to own directory
