+++ web/modules/custom/server_general/src/ThemeTrait/CarouselThemeTrait.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Drupal\server_general\ThemeTrait;
 
+use Drupal\server_general\ThemeTrait\Enum\BackgroundColorEnum;
+
 /**
  * Helper methods for rendering Carousel elements.
  */
