+++ web/modules/custom/server_general/src/ThemeTrait/ElementLayoutThemeTrait.php
@@ -6,6 +6,9 @@ namespace Drupal\server_general\ThemeTrait;
 
 use Drupal\Core\Link;
 use Drupal\Core\Url;
+use Drupal\server_general\ThemeTrait\Enum\BackgroundColorEnum;
+use Drupal\server_general\ThemeTrait\Enum\TextColorEnum;
+use Drupal\server_general\ThemeTrait\Enum\WidthEnum;
 
 /**
  * Helper methods to build Page layouts.
@@ -74,7 +77,7 @@ trait ElementLayoutThemeTrait {
    *   The body render array. Maybe empty.
    * @param array $items
    *   The items render array.
-   * @param BackgroundColorEnum $bg_color
+   * @param \Drupal\server_general\ThemeTrait\Enum\BackgroundColorEnum $bg_color
    *   The background color. See
    *   ElementWrapThemeTrait::wrapContainerWide for the allowed values.
    *
