+++ web/modules/custom/server_general/src/ThemeTrait/ElementNodeNewsThemeTrait.php
@@ -6,6 +6,8 @@ namespace Drupal\server_general\ThemeTrait;
 
 use Drupal\intl_date\IntlDate;
 use Drupal\server_general\EntityDateTrait;
+use Drupal\server_general\ThemeTrait\Enum\FontSizeEnum;
+use Drupal\server_general\ThemeTrait\Enum\WidthEnum;
 
 /**
  * Helper method for building the Node news element.
