+++ web/modules/custom/server_general/src/ThemeTrait/Enum/FontWeightEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\server_general\ThemeTrait;
+namespace Drupal\server_general\ThemeTrait\Enum;
 
 /**
  * Enum for font weight options used in theme wrappers.
