+++ web/modules/custom/server_general/src/ThemeTrait/Enum/UnderlineEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\server_general\ThemeTrait;
+namespace Drupal\server_general\ThemeTrait\Enum;
 
 /**
  * Enum for underline options used in theme wrappers.
