+++ web/modules/custom/server_general/src/ThemeTrait/PeopleTeasersThemeTrait.php
@@ -4,6 +4,11 @@ declare(strict_types=1);
 
 namespace Drupal\server_general\ThemeTrait;
 
+use Drupal\server_general\ThemeTrait\Enum\AlignmentEnum;
+use Drupal\server_general\ThemeTrait\Enum\FontSizeEnum;
+use Drupal\server_general\ThemeTrait\Enum\FontWeightEnum;
+use Drupal\server_general\ThemeTrait\Enum\TextColorEnum;
+
 /**
  * Helper methods for rendering People/Person Teaser elements.
  */
