+++ web/modules/custom/server_general/src/ThemeTrait/QuoteThemeTrait.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Drupal\server_general\ThemeTrait;
 
+use Drupal\server_general\ThemeTrait\Enum\FontSizeEnum;
+use Drupal\server_general\ThemeTrait\Enum\TextColorEnum;
+
 /**
  * Helper methods for rendering Quote elements.
  */
