+++ web/modules/custom/server_general/src/ThemeTrait/TitleAndLabelsThemeTrait.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Drupal\server_general\ThemeTrait;
 
+use Drupal\server_general\ThemeTrait\Enum\HtmlTagEnum;
+
 /**
  * Helper method for building Title and labels of a content.
  */
