+++ web/modules/custom/server_style_guide/src/Controller/StyleGuideController.php
@@ -9,7 +9,7 @@ use Drupal\pluggable_entity_view_builder\BuildFieldTrait;
 use Drupal\server_general\ThemeTrait\AccordionThemeTrait;
 use Drupal\server_general\ThemeTrait\ButtonThemeTrait;
 use Drupal\server_general\ThemeTrait\CardThemeTrait;
-use Drupal\server_general\ThemeTrait\ColorEnum;
+use Drupal\server_general\ThemeTrait\Enum\ColorEnum;
 use Drupal\server_general\ThemeTrait\CarouselThemeTrait;
 use Drupal\server_general\ThemeTrait\CtaThemeTrait;
 use Drupal\server_general\ThemeTrait\DocumentsThemeTrait;
@@ -18,12 +18,12 @@ use Drupal\server_general\ThemeTrait\ElementMediaThemeTrait;
 use Drupal\server_general\ThemeTrait\ElementNodeNewsThemeTrait;
 use Drupal\server_general\ThemeTrait\ElementWrapThemeTrait;
 use Drupal\server_general\ThemeTrait\ExpandingTextThemeTrait;
-use Drupal\server_general\ThemeTrait\FontSizeEnum;
-use Drupal\server_general\ThemeTrait\FontWeightEnum;
+use Drupal\server_general\ThemeTrait\Enum\FontSizeEnum;
+use Drupal\server_general\ThemeTrait\Enum\FontWeightEnum;
 use Drupal\server_general\ThemeTrait\HeroThemeTrait;
-use Drupal\server_general\ThemeTrait\HtmlTagEnum;
+use Drupal\server_general\ThemeTrait\Enum\HtmlTagEnum;
 use Drupal\server_general\ThemeTrait\InfoCardThemeTrait;
-use Drupal\server_general\ThemeTrait\LineClampEnum;
+use Drupal\server_general\ThemeTrait\Enum\LineClampEnum;
 use Drupal\server_general\ThemeTrait\LinkThemeTrait;
 use Drupal\server_general\ThemeTrait\NewsTeasersThemeTrait;
 use Drupal\server_general\ThemeTrait\PeopleTeasersThemeTrait;
@@ -33,7 +33,7 @@ use Drupal\server_general\ThemeTrait\SearchThemeTrait;
 use Drupal\server_general\ThemeTrait\SocialShareThemeTrait;
 use Drupal\server_general\ThemeTrait\TagThemeTrait;
 use Drupal\server_general\ThemeTrait\TitleAndLabelsThemeTrait;
-use Drupal\server_general\ThemeTrait\UnderlineEnum;
+use Drupal\server_general\ThemeTrait\Enum\UnderlineEnum;
 use Drupal\server_general\WebformTrait;
 use Drupal\server_style_guide\ThemeTrait\StyleGuideElementWrapThemeTrait;
 use Symfony\Component\DependencyInjection\ContainerInterface;
